allow --include_path to be repeated in tt-render
At line 80 special care was taken to set {include_path} to an array ref,
yet at line 83 it is given the same accessor as the other config options,
so that the arrayref will be overwritten by a string value when using tt-render.
This commit gives {include_path} its special accessor at line 169
which ensures that

    *   If the new value is an arrayref it is assigned to {include_path}.

    *   If the old value of {include_path} is not an arrayref
        it is split on the current {delimiter}
        and the resulting array is assigned to {include_path}.

    *   If the new value is not an arrayref
        it is split on the current {delimiter}
        and the resulting array is pushed onto the existing arrayref.

N.B. I'm not sure how to update the documentation to reflect this, since
it seems the POD is generated from the file at doc/Template/Toolkit/Simple.swim.
If you have any pointers ill add a note.
I don't know if this needs any test(s). It's pretty straightforward.